refactor: extract channel store admin command helper
